requirements/custom_models.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/custom_models.in
#
aiobotocore[awscli,boto3]==1.3.3
    # via
    #   -r requirements/common.in
    #   s3fs
aiohttp==3.7.4.post0
    # via aiobotocore
aioitertools==0.8.0
    # via aiobotocore
ansiwrap==0.8.4
    # via papermill
appdirs==1.4.4
    # via black
asn1crypto==1.4.0
    # via scramp
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   jsonschema
awscli==1.19.106
    # via aiobotocore
awswrangler==2.10.0
    # via -r requirements/common.in
bcrypt==3.2.0
    # via paramiko
beautifulsoup4==4.9.3
    # via redshift-connector
black==21.7b0
    # via papermill
blosc==1.10.2
    # via -r requirements/common.in
bokeh==2.3.3
    # via dask
boto3==1.17.106
    # via
    #   aiobotocore
    #   awswrangler
    #   pyathena
    #   redshift-connector
    #   tentaclio
botocore==1.20.106
    # via
    #   aiobotocore
    #   awscli
    #   awswrangler
    #   boto3
    #   pyathena
    #   redshift-connector
    #   s3transfer
cachetools==4.2.2
    # via google-auth
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via
    #   bcrypt
    #   cryptography
    #   google-crc32c
    #   pynacl
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
click==8.0.1
    # via
    #   -r requirements/common.in
    #   black
    #   distributed
    #   papermill
    #   tentaclio
cloudpickle==2.0.0
    # via
    #   -r requirements/common.in
    #   dask
    #   distributed
colorama==0.4.3
    # via awscli
cramjam==2.3.2
    # via fastparquet
cryptography==3.4.7
    # via paramiko
cycler==0.10.0
    # via matplotlib
dask[complete]==2021.09.1
    # via
    #   -r requirements/common.in
    #   distributed
dataclasses==0.6
    # via tentaclio
distributed==2021.09.1
    # via dask
docutils==0.15.2
    # via awscli
entrypoints==0.3
    # via papermill
et-xmlfile==1.1.0
    # via openpyxl
fastparquet==0.7.1
    # via -r requirements/common.in
freezegun==1.1.0
    # via -r requirements/common.in
fsspec==2021.7.0
    # via
    #   dask
    #   fastparquet
    #   s3fs
future==0.18.2
    # via pyathena
google-api-core==1.31.2
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.15.0
    # via tentaclio
google-auth==1.35.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
    #   gspread
google-auth-httplib2==0.1.0
    # via
    #   google-api-python-client
    #   tentaclio
google-auth-oauthlib==0.4.5
    # via
    #   gspread
    #   tentaclio
google-cloud-core==1.7.2
    # via google-cloud-storage
google-cloud-storage==1.42.0
    # via tentaclio
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==1.3.3
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
gspread==4.0.1
    # via -r requirements/common.in
heapdict==1.0.1
    # via zict
httplib2==0.19.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==2.10
    # via
    #   requests
    #   yarl
ipython-genutils==0.2.0
    # via
    #   nbformat
    #   traitlets
jinja2==3.0.1
    # via
    #   bokeh
    #   dask
    #   distributed
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via nbformat
jupyter-client==6.1.12
    # via nbclient
jupyter-core==4.7.1
    # via
    #   jupyter-client
    #   nbformat
kiwisolver==1.3.1
    # via matplotlib
kotsu==0.0.1a1
    # via -r requirements/custom_models.in
locket==0.2.1
    # via partd
lxml==4.6.3
    # via redshift-connector
lz4==3.1.3
    # via -r requirements/common.in
markupsafe==2.0.1
    # via jinja2
matplotlib==3.4.3
    # via -r requirements/common.in
msgpack==1.0.2
    # via
    #   -r requirements/common.in
    #   distributed
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via black
nbclient==0.5.4
    # via papermill
nbformat==5.1.3
    # via
    #   nbclient
    #   papermill
nest-asyncio==1.5.1
    # via nbclient
numpy==1.21.2
    # via
    #   awswrangler
    #   bokeh
    #   dask
    #   fastparquet
    #   matplotlib
    #   pandas
    #   pyarrow
oauthlib==3.1.1
    # via requests-oauthlib
openpyxl==3.0.7
    # via awswrangler
packaging==21.0
    # via
    #   bokeh
    #   dask
    #   google-api-core
    #   redshift-connector
pandas==1.1.3
    # via
    #   awswrangler
    #   dask
    #   fastparquet
    #   tentaclio
papermill==2.3.3
    # via -r requirements/common.in
paramiko==2.7.2
    # via pysftp
partd==1.2.0
    # via dask
pathspec==0.9.0
    # via black
pg8000==1.20.0
    # via awswrangler
pillow==8.3.1
    # via
    #   bokeh
    #   matplotlib
protobuf==3.17.3
    # via
    #   google-api-core
    #   googleapis-common-protos
psutil==5.8.0
    # via distributed
psycopg2-binary==2.9.1
    # via tentaclio
pyarrow==4.0.1
    # via
    #   -r requirements/common.in
    #   awswrangler
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyathena==1.10.7
    # via tentaclio
pycparser==2.20
    # via cffi
pymysql==1.0.2
    # via awswrangler
pynacl==1.4.0
    # via paramiko
pyodbc==4.0.31
    # via tentaclio
pyparsing==2.4.7
    # via
    #   httplib2
    #   matplotlib
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
pysftp==0.2.9
    # via tentaclio
python-dateutil==2.8.2
    # via
    #   bokeh
    #   botocore
    #   freezegun
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2021.1
    # via
    #   google-api-core
    #   pandas
    #   redshift-connector
pyyaml==5.4.1
    # via
    #   -r requirements/common.in
    #   awscli
    #   bokeh
    #   dask
    #   distributed
    #   papermill
    #   tentaclio
pyzmq==22.2.1
    # via jupyter-client
redshift-connector==2.0.885
    # via awswrangler
regex==2021.8.3
    # via black
requests==2.25.1
    # via
    #   google-api-core
    #   google-cloud-storage
    #   papermill
    #   redshift-connector
    #   requests-oauthlib
    #   tentaclio
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.7.2
    # via
    #   awscli
    #   google-auth
s3fs==2021.7.0
    # via -r requirements/common.in
s3transfer==0.4.2
    # via
    #   awscli
    #   boto3
scramp==1.4.0
    # via
    #   pg8000
    #   redshift-connector
six==1.16.0
    # via
    #   bcrypt
    #   cycler
    #   google-api-core
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-resumable-media
    #   jsonschema
    #   protobuf
    #   pynacl
    #   python-dateutil
    #   thrift
sortedcontainers==2.4.0
    # via distributed
soupsieve==2.2.1
    # via beautifulsoup4
sqlalchemy==1.3.24
    # via tentaclio
tblib==1.7.0
    # via distributed
tenacity==8.0.1
    # via
    #   papermill
    #   pyathena
tentaclio==0.0.15
    # via -r requirements/common.in
textwrap3==0.9.2
    # via ansiwrap
thrift==0.13.0
    # via fastparquet
tomli==1.2.1
    # via black
toolz==0.11.1
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.1
    # via
    #   bokeh
    #   distributed
    #   jupyter-client
tqdm==4.62.1
    # via papermill
traitlets==5.0.5
    # via
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbformat
types-pyyaml==6.0.1
    # via -r requirements/common.in
typing-extensions==3.10.0.0
    # via
    #   aiohttp
    #   bokeh
    #   tentaclio
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.6
    # via
    #   botocore
    #   requests
    #   tentaclio
wrapt==1.12.1
    # via aiobotocore
yarl==1.6.3
    # via aiohttp
zict==2.0.0
    # via distributed

# The following packages are considered to be unsafe in a requirements file:
# setuptools
